Bufferedoutput/input stream
File I/O

+++ namoo.IO/src/namoo/IO/FileCopyExample.java
@@ -1,5 +1,6 @@
 package namoo.IO;
 
+import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
@@ -13,18 +14,28 @@ public class FileCopyExample {
 		String srcpath = "D:\\웹개발_이규헌\\workspace\\namoo.IO\\example.txt";
 		String destPath = "D:\\웹개발_이규헌\\workspace\\namoo.IO\\example2.txt";
 		
-		InputStream in = new FileInputStream(srcpath);
-		OutputStream out = new FileOutputStream(destPath);
-		byte[] buffer = new byte[4*1024];
-		int count = 0;
-		
-		while((count = in.read(buffer)) != -1) {
-			out.write(buffer, 0, count); //맨 마지막에 남은 바이트만큼만 써야하기때문에 offset을 사용하여 조건 범위를 설정해줌
+		File srcfile = new File(srcpath);
+		if(srcfile.exists()){
+			InputStream in = new FileInputStream(srcpath);
+			OutputStream out = new FileOutputStream(destPath);
+			byte[] buffer = new byte[4*1024];
+			int count = 0;
 			
+			while((count = in.read(buffer)) != -1) {
+				out.write(buffer, 0, count); //맨 마지막에 남은 바이트만큼만 써야하기때문에 offset을 사용하여 조건 범위를 설정해줌
+				
+				
+			}
 			
+			System.out.println(srcfile.length() + "byte 파일 복사완료");
+			if(out != null)out.close();
+			if(in != null)in.close();
+			
+		}else {
+			System.out.println("복사 불가");
 		}
-		System.out.println("파일 복사완료");
-		if(out != null)out.close();
-		if(in != null)in.close();
+		
+		
+		
 	}
 }
